pyproject.toml
[project]
name = "graphloop-backend"
version = "0.1.0"
description = "Backend for GraphLoop project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "deepagents>=0.0.5",
    "e2b-code-interpreter>=2.2.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black",
    "ruff",
    "langgraph-cli[inmem]>=0.4.3",
]
